BUG: Register parameterNodeWrapper plugins in corresponding module
This commit fixes the following tests by ensuring parameterNodeWrapper
guiConnectors specific to modules are registered in the corresponding
modules.

  py_nowarning_mainwindow_noloadableTest
  py_nowarning_mainwindow_nocli_noloadableTest
  py_nowarning_nomainwindow_noloadableTest
  py_nowarning_nomainwindow_nocli_noloadableTest
  py_nomainwindow_SlicerOptionModulesToIgnoreTest

Note that since the test associated with `qMRMLSubjectHierarchyTreeViewToNodeConnector`
is still implemented in `py_nomainwindow_test_slicer_parameter_node_wrapper_guis`,
it is required to have the `SubjectHierarchy` python classes imported by
removing the `--disable-scripted-loadable-modules` argument.
