+++ Applications/SlicerApp/Testing/Python/CMakeLists.txt
@@ -310,7 +310,7 @@ slicer_add_python_unittest(
 
 slicer_add_python_unittest(
   SCRIPT ${Slicer_SOURCE_DIR}/Base/Python/slicer/tests/test_slicer_parameter_node_wrapper_guis.py
-  SLICER_ARGS --no-main-window --disable-cli-modules --disable-scripted-loadable-modules
+  SLICER_ARGS --no-main-window --disable-cli-modules
   TESTNAME_PREFIX nomainwindow_
   )
