+++ Modules/Loadable/SubjectHierarchy/Widgets/Python/parameterNodeWrapper/__init__.py
@@ -1,0 +1,1 @@
+from .guiConnectors import *
